# doc/agency.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "agency/agency.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/v1/account/login": {
      "post": {
        "operationId": "AccountLogin",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/agencyAccountLoginResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/agencyAccountLoginRequest"
            }
          }
        ],
        "tags": [
          "BlueAgency"
        ]
      }
    },
...
  },
  "definitions": {
...
    "agencyAccountLoginRequest": {
      "type": "object",
      "properties": {
        "phone": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "agencyAccountLoginResponse": {
      "type": "object",
      "properties": {
        "user_id": {
          "type": "string"
        },
        "access_token": {
          "type": "string"
        },
        "user_info": {
          "$ref": "#/definitions/agencyAgencyStaff"
        }
      }
    },
...
    "agencyAgencyStaff": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "address": {
          "type": "string"
        }
      }
    },
...
  }
}